Identify Car brand using Scrollview
Clicking on the Identify the Car Make a button, it will be displayed to the user one random car image picked randomly (and NOT in the same random sequence every time the application restarts) from the set of all the car images that your application contains. The screen will also display a dropdown list with ALL the names of the car makes that your application contains (e.g. BMW, Ford, etc.) and a button labeled Identify. The user will use the list to select the car make that the car image corresponds (b) As soon as the user submits the answer, the message CORRECT! (in green color) or the message WRONG! (in red color) appears on the screen with the name of the correct car make in Yellow color, depending on whether the answer given is correct or incorrect respectively.  the label of the “Identify” button will be changed to display “Next”. The user should click the “Next” button to advance to another screen presenting him with a new random car image and giving him the chance to play again. Every time that this option is chosen a different image will be displayed.