Advance feature and points
 Clicking on the Advanced Level button, it will be displayed to the user 3 different unique random car images picked from all the car images in the application together with 3 textboxes and a Submit button. The aim of the user in this level is to guess the car manufacturer names of all the 3 displayed cars using the corresponding 3 textboxes. The name of the car manufacturer corresponding to the first image will be typed in the first textbox, the name of the manufacturer corresponding to the second image will be typed in the second textbox and the name of the manufacturer corresponding to the third image will be typed in the third textbox. The user will be using the “Submit” button to submit his answers. Once he clicks on that button the answers are checked. If any of the 3 textboxes contain the correct car make name then that textbox becomes uneditable (i.e. the user cannot change any characters in it) and in GREEN color. Any textboxes containing an incorrect guess remain editable but with a RED color (either the textbox background becomes red or the characters face becomes red) . If the user gets all the 3 manufacturers correct, then the message CORRECT! (in green colour) appears, the label of the “Submit” button becomes “Next” and clicking on the button leads the users to a new game with 3 new images being displayed.

The user is allowed only 3 incorrect attempts. A single attempt is considered complete every time that the “Submit” button is clicked. After 3 incorrect attempts (i.e. the user has clicked the “Submit” button 3 times and at least 1 of the 3 car make guesses is incorrect) the message WRONG! (in red color) appears on the screen with the names of the correct manufacturers in Yellow colour displayed just below the incorrect answer(s). (If 1 or 2 guesses are correct the “Yellow” name does not appear below the correct answers.) After the 3 incorrect attempts the label of the “Submit” button becomes “Next” and clicking on the button leads the users to a new game with 3 new images being displayed. Implement a score for the user which is displayed on the top right of the screen. The user gets 1 point for every image he guesses correctly, i.e. if in the first screen he guesses all 3 car makes correctly he gets 3 points. If after 3 attempts in a screen, he guesses correctly only 2 manufacturers out of the 3, he gets 2 points and so on.